requirements.txt
python-telegram-bot==20.1
aiohttp==3.8.4

flake8>=6.0.0
isort>=5.12.0
black>=23.1.0
mypy>=1.1.1
